[C] Refactor GlobalStyles and export style utils
For git: "Resolves 52"

For JIRA: "EPO-7924 #IN-REVIEW #comment export style utils and refactor GlobalStyles"
