+++ packages/epo-react-lib/src/form/FormButtons/styles.ts
@@ -1,5 +1,5 @@
 import styled from "styled-components";
-import { layoutGrid } from "@/styles/globalStyles";
+import { layoutGrid } from "@/styles/utils";
 
 export const FormButtons = styled.div`
   ${layoutGrid(2)}
