+++ packages/epo-react-lib/src/index.ts
@@ -1,10 +1,8 @@
 // styled-components
+export { tokens } from "@/styles/abstracts";
+export * from "@/styles/utils";
 
-export {
-  tokens,
-  fluidScale,
-  default as GlobalStyles,
-} from "@/styles/globalStyles";
+export { default as GlobalStyles } from "@/styles/globalStyles";
 
 // i18n
 export { default as localeStrings } from "@/assets/locales";
