+++ packages/epo-react-lib/src/styles/abstracts/index.ts
@@ -1,0 +1,2 @@
+export { default as palette } from "./palette";
+export * from "./tokens";
