+++ packages/epo-react-lib/src/styles/abstracts/palette.ts
@@ -1,0 +1,41 @@
+const palette: { [key: string]: string } = {
+  white: "#ffffff",
+  offWhite: "#e9e9e9",
+  neutral02: "#fafafa",
+  neutral08: "#ebebeb",
+  neutral10: "#f5f5f5",
+  neutral15: "#e3e4e8",
+  neutral20: "#dce0e3",
+  neutral30: "#d3d4d5",
+  neutral40: "#9E9EA3",
+  neutral60: "#6A6E6E",
+  neutral80: "#404040",
+  neutral90: "#2b2e34",
+  neutral95: "#1F2121",
+  black: "#000000",
+  turquoise05: "#EDFFFE",
+  turqouise07: "#EDFFFE",
+  turquoise10: "#D9F7F6",
+  turquoise20: "#B2F2EF",
+  turquoise50: "#00bebf",
+  turquoise55: "#009fa1",
+  turquoise60: "#078b8c",
+  turquoise70: "#058b8c",
+  turquoise80: "#117273",
+  turquoise85: "#12726c",
+  turquoise90: "#0c4a4c",
+  orange02: "#FFFDFA",
+  orange04: "#FFF8F0",
+  orange05: "#FFF9F2",
+  orange10: "#FFEDD9",
+  orange20: "#FFE7CC",
+  orange55: "#FAB364",
+  blue10: "#eef5fb",
+  blue20: "#C7D4F4",
+  green05: "#D7E6CF",
+  red: "#cf4040",
+  red20: "#f2c3c0",
+  red40: "#FF8489",
+};
+
+export default palette;
