+++ packages/epo-widget-lib/package.json
@@ -64,7 +64,7 @@
     "react-i18next": "^12.2.0"
   },
   "dependencies": {
-    "@rubin-epo/epo-react-lib": "^1.1.0",
+    "@rubin-epo/epo-react-lib": "^1.1.5",
     "styled-components": "^5.3.9",
     "use-resize-observer": "^9.1.0"
   }
